Bit more cleanup
